Reorganized spec files.
