+++ .travis.yml
@@ -1,0 +1,2 @@
+env:
+  - TRAVIS=true
