+++ spec/spec_helper.rb
@@ -1,3 +1,13 @@
-$LOAD_PATH.unshift(File.expand_path("../../lib", __FILE__))
-require 'coveralls'
-Coveralls.wear!
+GEM_ROOT = File.expand_path("../../", __FILE__)
+$:.unshift File.join(GEM_ROOT, "lib")
+
+if ENV["TRAVIS"]
+  require 'coveralls'
+  Coveralls.wear!
+else
+  require 'simplecov'
+  SimpleCov.start
+end
+
+require 'naught'
+Dir[File.join(GEM_ROOT, "spec", "support", "**/*.rb")].each { |f| require f }
