+++ spec/support/convertable_null.rb
@@ -1,0 +1,4 @@
+ConvertableNull = Naught.build do |b|
+  b.null_equivalents << ""
+  b.traceable
+end
